refactor: revert GalleryImage.vue rename back to ProfileImage.vue
The earlier rename of ProfileImage.vue -> GalleryImage.vue (commit
bc4dbbd5) shipped consistent naming with the broader gallery
abstraction but ballooned the diff vs main without semantic gain — the
component still primarily renders profile images. Renames the file
back, along with its test (GalleryImage.spec.ts -> ProfileImage.spec.ts)
and all 12 import/template references.

Docs under docs/superpowers/ are intentionally not touched: their
remaining GalleryImage references are all directional ("rename X -> Y")
historical instructions that would be corrupted by a blanket rename.
