+++ apps/frontend/src/features/browse/views/__tests__/BrowseProfiles.spec.ts
@@ -219,7 +219,7 @@ describe('BrowseProfiles view', () => {
           BContainer,
           InviteCtaShareDialog,
           NotificationDot: { template: '<span><slot /></span>' },
-          GalleryImage: true,
+          ProfileImage: true,
           Teleport: true,
         },
         mocks: { $t: (k: string) => k },
