+++ apps/frontend/src/features/images/__tests__/ProfileThumbnail.spec.ts
@@ -1,12 +1,12 @@
 import { mount } from '@vue/test-utils'
 import { describe, it, expect, vi } from 'vitest'
 
-vi.mock('../components/GalleryImage.vue', () => ({ default: { template: '<div class="img" />' } }))
+vi.mock('../components/ProfileImage.vue', () => ({ default: { template: '<div class="img" />' } }))
 
 import ProfileThumbnail from '../components/ProfileThumbnail.vue'
 
 describe('ProfileThumbnail', () => {
-  it('renders GalleryImage', () => {
+  it('renders ProfileImage', () => {
     const wrapper = mount(ProfileThumbnail, { props: { profile: {} as any } })
     expect(wrapper.find('.img').exists()).toBe(true)
   })
