+++ apps/frontend/src/features/images/components/ProfileImage.vue
